Most of the changes here are trivial, but the white space changes would be harder to undo than to do over.
I changed all groups of 8 spaces to tabs, then all tabs to two spaces so more of the device tree fits on the page.  It could have been three or possibly four, but the largest indents I found were 6 tabs, so 4 is a lot of the space on the page.

Signed-off-by: Myles Watson <mylesgw@gmail.com>
Acked-by: Ronald G. Minnich <rminnich@gmail.com>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@4543 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1